pass client response header iterable to _debug
Httplib2 response object is an iterable of response headers. Urllib3 changes this and the response headers are only available from `response.headers`. The `_debug` method was written for Httplib2 and expects to be able to check for the `content-type` header in the first argument, so the headers must be passed instead of the response.